moves display dinosaur into its own folder
