+++ scripts/components/displayDinosaur/displayDinosaur.styles.scss
@@ -1,4 +1,4 @@
-  
+ @import "../../../styles/abstracts/mixins";
   .modal-container {
     height: 75vh;
     width: 75vw;
