+++ styles/main.scss
@@ -17,7 +17,7 @@
 @import "components/map";
 @import "components/heart";
 @import "components/dino-list";
-@import "components/dino-modal";
+
 
 @import "layout/grid";
 @import "layout/header";
